chore: 🐋 Tab type 파일 분리 및 적용
